+++ bin/check-ec2-network.rb
@@ -12,7 +12,7 @@
 #   Linux
 #
 # DEPENDENCIES:
-#   gem: aws-sdk
+#   gem: aws-sdk-v1
 #   gem: sensu-plugin
 #
 # USAGE:
@@ -32,20 +32,23 @@ require 'sensu-plugin/check/cli'
 require 'aws-sdk-v1'
 
 class CheckEc2Network < Sensu::Plugin::Check::CLI
-  option :access_key_id,
-         short:       '-k N',
-         long:        '--access-key-id ID',
-         description: 'AWS access key ID'
+  option :aws_access_key,
+         short:       '-a AWS_ACCESS_KEY',
+         long:        '--aws-access-key AWS_ACCESS_KEY',
+         description: "AWS Access Key. Either set ENV['AWS_ACCESS_KEY'] or provide it as an option",
+         default:     ENV['AWS_ACCESS_KEY']
 
-  option :secret_access_key,
-         short:       '-s N',
-         long:        '--secret-access-key KEY',
-         description: 'AWS secret access key'
+  option :aws_secret_access_key,
+         short:       '-k AWS_SECRET_KEY',
+         long:        '--aws-secret-access-key AWS_SECRET_KEY',
+         description: "AWS Secret Access Key. Either set ENV['AWS_SECRET_KEY'] or provide it as an option",
+         default:     ENV['AWS_SECRET_KEY']
 
-  option :region,
-         short:       '-r R',
-         long:        '--region REGION',
-         description: 'AWS region'
+  option :aws_region,
+         short:       '-r AWS_REGION',
+         long:        '--aws-region REGION',
+         description: 'AWS Region (defaults to us-east-1).',
+         default:     'us-east-1'
 
   option :instance_id,
          short:       '-i instance-id',
@@ -77,10 +80,10 @@ class CheckEc2Network < Sensu::Plugin::Check::CLI
   end
 
   def aws_config
-    hash = {}
-    hash.update access_key_id: config[:access_key_id], secret_access_key: config[:secret_access_key] if config[:access_key_id] && config[:secret_access_key]
-    hash.update region: config[:region] if config[:region]
-    hash
+    { access_key_id: config[:aws_access_key],
+      secret_access_key: config[:aws_secret_access_key],
+      region: config[:aws_region]
+    }
   end
 
   def ec2
@@ -106,10 +109,7 @@ class CheckEc2Network < Sensu::Plugin::Check::CLI
 
   def latest_value(metric)
     value = metric.statistics(statistics_options.merge unit: 'Bytes')
-    # #YELLOW
-    unless value.datapoints[0].nil? # rubocop:disable IfUnlessModifier, GuardClause
-      value.datapoints[0][:average].to_f
-    end
+    value.datapoints[0][:average].to_f unless value.datapoints[0].nil?
   end
 
   def check_metric(instance)