+++ bin/check-rds-events.rb
@@ -16,7 +16,7 @@
 #   Linux
 #
 # DEPENDENCIES:
-#   gem: aws-sdk
+#   gem: aws-sdk-v1
 #   gem: sensu-plugin
 #
 # USAGE:
@@ -35,28 +35,41 @@ require 'aws-sdk-v1'
 
 class CheckRDSEvents < Sensu::Plugin::Check::CLI
   option :aws_access_key,
-         short: '-a AWS_ACCESS_KEY',
-         long: '--aws-access-key AWS_ACCESS_KEY',
-         description: "AWS Access Key. Either set ENV['AWS_ACCESS_KEY_ID'] or provide it as an option",
-         default: ENV['AWS_ACCESS_KEY_ID']
+         short:       '-a AWS_ACCESS_KEY',
+         long:        '--aws-access-key AWS_ACCESS_KEY',
+         description: "AWS Access Key. Either set ENV['AWS_ACCESS_KEY'] or provide it as an option",
+         default:     ENV['AWS_ACCESS_KEY']
 
   option :aws_secret_access_key,
-         short: '-s AWS_SECRET_ACCESS_KEY',
-         long: '--aws-secret-access-key AWS_SECRET_ACCESS_KEY',
-         description: "AWS Secret Access Key. Either set ENV['AWS_SECRET_ACCESS_KEY'] or provide it as an option",
-         default: ENV['AWS_SECRET_ACCESS_KEY']
+         short:       '-k AWS_SECRET_KEY',
+         long:        '--aws-secret-access-key AWS_SECRET_KEY',
+         description: "AWS Secret Access Key. Either set ENV['AWS_SECRET_KEY'] or provide it as an option",
+         default:     ENV['AWS_SECRET_KEY']
 
   option :aws_region,
-         short: '-r AWS_REGION',
-         long: '--aws-region REGION',
-         description: 'AWS Region (such as eu-west-1).',
-         default: 'us-east-1'
+         short:       '-r AWS_REGION',
+         long:        '--aws-region REGION',
+         description: 'AWS Region (defaults to us-east-1).',
+         default:     'us-east-1'
 
-  def run # rubocop:disable AbcSize
-    rds = AWS::RDS::Client.new(
-      access_key_id: config[:aws_access_key],
+  def aws_config
+    { access_key_id: config[:aws_access_key],
       secret_access_key: config[:aws_secret_access_key],
-      region: config[:aws_region])
+      region: config[:aws_region]
+    }
+  end
+
+  def run
+    clusters = maint_clusters
+    if clusters.empty?
+      ok
+    else
+      critical("Clusters w/ critical events: #{clusters.join(',')}")
+    end
+  end
+
+  def maint_clusters
+    rds = AWS::RDS::Client.new aws_config
 
     begin
       # fetch all clusters identifiers
@@ -74,11 +87,6 @@ class CheckRDSEvents < Sensu::Plugin::Check::CLI
     rescue => e
       unknown "An error occurred processing AWS RDS API: #{e.message}"
     end
-
-    if maint_clusters.empty?
-      ok
-    else
-      critical("Clusters w/ critical events: #{maint_clusters.join(',')}")
-    end
+    maint_clusters
   end
 end