+++ bin/metrics-elb-full.rb
@@ -12,6 +12,7 @@
 #   Linux
 #
 # DEPENDENCIES:
+#   gem: aws-sdk-v1
 #   gem: sensu-plugin
 #
 # USAGE:
@@ -37,9 +38,10 @@ require 'aws-sdk-v1'
 
 class ELBMetrics < Sensu::Plugin::Metric::CLI::Graphite
   option :elbname,
-         description: 'Name of the Elastic Load Balancer',
+         description: 'Name of the Elastic Load Balancer (separate with commas for multiple ELBs)',
          short: '-n ELB_NAME',
-         long: '--name ELB_NAME'
+         long: '--name ELB_NAME',
+         required: true
 
   option :scheme,
          description: 'Metric naming scheme, text to prepend to metric',
@@ -58,27 +60,25 @@ class ELBMetrics < Sensu::Plugin::Metric::CLI::Graphite
          short: '-a AWS_ACCESS_KEY',
          long: '--aws-access-key AWS_ACCESS_KEY',
          description: "AWS Access Key. Either set ENV['AWS_ACCESS_KEY'] or provide it as an option",
-         required: true,
          default: ENV['AWS_ACCESS_KEY']
 
   option :aws_secret_access_key,
          short: '-k AWS_SECRET_KEY',
          long: '--aws-secret-access-key AWS_SECRET_KEY',
          description: "AWS Secret Access Key. Either set ENV['AWS_SECRET_KEY'] or provide it as an option",
-         required: true,
          default: ENV['AWS_SECRET_KEY']
 
   option :aws_region,
          short: '-r AWS_REGION',
          long: '--aws-region REGION',
-         description: 'AWS Region (such as eu-west-1).',
+         description: 'AWS Region (defaults to us-east-1).',
          default: 'us-east-1'
 
   def aws_config
-    hash = {}
-    hash.update access_key_id: config[:access_key_id], secret_access_key: config[:secret_access_key] if config[:access_key_id] && config[:secret_access_key]
-    hash.update region: config[:aws_region]
-    hash
+    { access_key_id: config[:aws_access_key],
+      secret_access_key: config[:aws_secret_access_key],
+      region: config[:aws_region]
+    }
   end
 
   def run
@@ -109,7 +109,7 @@ class ELBMetrics < Sensu::Plugin::Metric::CLI::Graphite
         'dimensions' => [
           {
             'name' => 'LoadBalancerName',
-            'value' => config[:elbname]
+            'value' => nil
           }
         ],
         'start_time' => st.iso8601,
@@ -118,13 +118,15 @@ class ELBMetrics < Sensu::Plugin::Metric::CLI::Graphite
       }
 
       result = {}
-      graphitepath = config[:scheme]
 
-      config[:elbname].split(' ').each do |elbname|
+      config[:elbname].split(',').each do |elbname|
+        if config[:scheme] == ''
+          graphitepath = "#{elbname}"
+        else
+          graphitepath = "#{config[:scheme]}.#{elbname}"
+        end
+
         statistic_type.each do |key, value|
-          if config[:scheme] == ''
-            graphitepath = "#{config[:elbname]}.#{key.downcase}"
-          end
           options['metric_name'] = key
           options['dimensions'][0]['value'] = elbname
           options['statistics'] = [value]