+++ bin/metrics-elb.rb
@@ -1,6 +1,6 @@
 #! /usr/bin/env ruby
 #
-# elb-latency-metrics
+# elb-metrics
 #
 # DESCRIPTION:
 #   Gets latency metrics from CloudWatch and puts them in Graphite for longer term storage
@@ -12,6 +12,7 @@
 #   Linux
 #
 # DEPENDENCIES:
+#   gem: aws-sdk-v1
 #   gem: sensu-plugin
 #
 # USAGE:
@@ -39,7 +40,8 @@ class ELBMetrics < Sensu::Plugin::Metric::CLI::Graphite
   option :elbname,
          description: 'Name of the Elastic Load Balancer',
          short: '-n ELB_NAME',
-         long: '--name ELB_NAME'
+         long: '--name ELB_NAME',
+         required: true
 
   option :scheme,
          description: 'Metric naming scheme, text to prepend to metric',
@@ -81,18 +83,17 @@ class ELBMetrics < Sensu::Plugin::Metric::CLI::Graphite
   option :aws_region,
          short: '-r AWS_REGION',
          long: '--aws-region REGION',
-         description: 'AWS Region (such as eu-west-1).',
+         description: 'AWS Region (defaults to us-east-1).',
          default: 'us-east-1'
 
   def aws_config
-    hash = {}
-    hash.update access_key_id: config[:access_key_id], secret_access_key: config[:secret_access_key] if config[:access_key_id] && config[:secret_access_key]
-    hash.update region: config[:aws_region]
-    hash
+    { access_key_id: config[:aws_access_key],
+      secret_access_key: config[:aws_secret_access_key],
+      region: config[:aws_region]
+    }
   end
 
   def run
-    statistic = ''
     if config[:statistic] == ''
       default_statistic_per_metric = {
         'Latency' => 'Average',